// web/projects/portal/src/app/vs-wizard/model/recommender/vs-object-recommendation.ts
/*
 * This file is part of StyleBI.
 * Copyright (C) 2024  InetSoft Technology
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
import { VSRecommendType } from "./vs-recommend-type";
import { VSSubType } from "./vs-sub-type";

export interface VSObjectRecommendation {
   type: VSRecommendType;
   subTypes: VSSubType[];
   selectedIndex: number;
}